WIP: Move code into source file
